requirements.txt
Django==5.1.1
psycopg2-binary
ffmpeg-python
